move the comparers to a separate file
