+++ DDDEastAnglia/Controllers/SpeakerController.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Web.Mvc;
 using DDDEastAnglia.DataAccess;
@@ -66,24 +65,6 @@ namespace DDDEastAnglia.Controllers
                     WebsiteUrl = userProfile.WebsiteUrl,
                     Sessions = userSessions
                 };
-        }
-    }
-
-    public class SpeakerDisplayModelComparer : IComparer<SpeakerDisplayModel>
-    {
-        public int Compare(SpeakerDisplayModel x, SpeakerDisplayModel y)
-        {
-            if (x.IsCurrentUser && !y.IsCurrentUser)
-            {
-                return -1;
-            }
-
-            if (!x.IsCurrentUser && y.IsCurrentUser)
-            {
-                return 1;
-            }
-
-            return string.Compare(x.Name, y.Name, StringComparison.InvariantCultureIgnoreCase);
         }
     }
 }
